requirements.txt
fake-bpy-module # type hints for bpy